Perform all PPU calculations with uints

+++ dotNES/CPU.Memory.cs
@@ -179,14 +179,14 @@ namespace dotNES
             }
         }
 
-        private void PerformDMA(int from)
+        private void PerformDMA(uint from)
         {
             //Console.WriteLine("OAM DMA");
             from <<= 8;
-            int OAMADDR = _emulator.PPU.F.OAMAddress;
-            for (int i = 0; i <= 0xFF; i++)
+            uint OAMADDR = _emulator.PPU.F.OAMAddress;
+            for (uint i = 0; i <= 0xFF; i++)
             {
-                _emulator.PPU.OAM[i] = (byte)ReadByte((uint)(from | ((i + OAMADDR) & 0xFF)));
+                _emulator.PPU.OAM[i] = (byte)ReadByte(from | ((i + OAMADDR) & 0xFF));
             }
             _cycle += 513;
         }
